Done with corrections

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,40 +1,33 @@
 #!/usr/bin/python3
-"""Delete all State objects with a name containing the letter `a` from
-database hbtn_0e_6_usa.
+"""
+This script deletes all State objects
+with a name containing the letter `a`
+from the database `hbtn_0e_6_usa`.
 """
 
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import Column
-from sqlalchemy import Integer
-from sqlalchemy import String
-
-Base = declarative_base()
-
-
-class State(Base):
-    """Class representing the `states` table.
-
-    Columns:
-        id (int): /NOT NULL/AUTO_INCREMENT/PRIMARY_KEY/
-        name (string): /VARCHAR(128)/NOT NULL/
-    """
-    __tablename__ = 'states'
-
-    id = Column(Integer, nullable=False, primary_key=True, autoincrement=True)
-    name = Column(String(128), nullable=False)
+from sys import argv
+from model_state import State, Base
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
 
 if __name__ == "__main__":
-    import sys
-    from sqlalchemy import create_engine
-    from sqlalchemy.orm import sessionmaker
+    """
+    Deletes State objects on the database.
+    """
 
-    engine = create_engine('mysql+mysqldb://'
-                           '{}:{}@localhost/{}'
-                           .format(sys.argv[1],
-                                   sys.argv[2],
-                                   sys.argv[3]))
+    db_url = "mysql+mysqldb://{}:{}@localhost:3306/{}".format(
+        argv[1], argv[2], argv[3])
+
+    engine = create_engine(db_url)
     Session = sessionmaker(bind=engine)
+
     session = Session()
-    for state in session.query(State).filter(State.name.like('%a%')):
-        session.delete(state)
+
+    states = session.query(State).filter(State.name.contains('a'))
+    if states is not None:
+        for state in states:
+            session.delete(state)
+
     session.commit()
+
+    session.close()
